Entity framework implemented

+++ Entities/Concrete/Car.cs
@@ -7,11 +7,11 @@ namespace Entities.Concrete
 {
     public class Car : IEntity
     {
-        public uint Id { get; set; }
-        public ushort BrandId { get; set; }
-        public uint ColorId { get; set; }
-        public ushort ModelYear { get; set; }
-        public uint DailyPrice { get; set; }
+        public int Id { get; set; }
+        public short BrandId { get; set; }
+        public short ColorId { get; set; }
+        public short ModelYear { get; set; }
+        public int DailyPrice { get; set; }
         public string Description { get; set; }
     }
 }
